requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
httpx==0.27.2
Pillow==11.0.0
easyocr==1.7.2
torch>=2.2.0
loguru==0.7.2

